requirements.txt
aiogram==3.13.1
chromadb==0.5.23
sentence-transformers==3.1.1
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.5.2